cargo clippy --fix

+++ src/world/atmosphere/mod.rs
@@ -36,7 +36,7 @@ fn daylight_cycle(
     timer.0.tick(time.delta());
 
     if timer.0.finished() {
-        let t = (time.elapsed_seconds_wrapped() as f32 / DAY_CYCLE_SECONDS)*2.0*PI;
+        let t = (time.elapsed_seconds_wrapped() / DAY_CYCLE_SECONDS)*2.0*PI;
         atmosphere.sun_position = Vec3::new(0.0, t.sin(), t.cos());
 
         if let Some((mut light_trans, mut directional)) = query.single_mut().into() {
